refactor parameter file
